Organize Python semantic model internals
